apps/bot/pyproject.toml
[project]
name = "smart-discord-bot"
version = "0.1.0"
description = "Discord.py Ingestion Worker"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "redis>=5.0.0",
    "celery>=5.4.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "qdrant-client>=1.12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
